testing/web-platform/tests/css/css-contain/content-visibility/content-visibility-026.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Content Visibility: containment added</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="content-visibility visible does not add containment">
<meta name="assert" content="content-visibility hidden adds containment">
<meta name="assert" content="content-visibility auto adds containment">
<meta name="assert" content="content-visibility auto visibility switches after rAF callbacks of the frame that produces the painted output">

<style>
.auto {
  content-visibility: auto;
}
.hidden {
  content-visibility: hidden;
}
.spacer {
  height: 10000px;
}
</style>
<div id="container"></div>
<div class=spacer></div>
<div id="offscreen_container"></div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
function setUp(container) {
  container.style = "";
  container.classList = "";
  assert_equals(getComputedStyle(container).contain, "none");
}

test(() => {
  setUp(container);
  container.classList.add("hidden");
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
}, "content-visibility: hidden adds contain: size layout style;");

async_test((t) => {
  setUp(container);
  setUp(offscreen_container);

  container.classList.add("auto");
  offscreen_container.classList.add("auto");
  // This will be determined to be on-screen immediately.
  t.step(() => assert_equals(getComputedStyle(container).contain, "size layout style paint", "initial onscreen"));
  t.step(() => assert_equals(getComputedStyle(offscreen_container).contain, "size layout style paint", "initial offscreen"));
  // Considering this as frame 1:
  // At frame 2, intersection observation will happen and determine that the container
  //   is not visible.
  // At frame 3, the container has size-containment, and intersection
  //   observations will mark this container as visible at the end of the frame
  // At frame 4, the container still has size-containment, because visiblility
  //   switch happens after rAF-callbacks have run.
  // At frame 5, the container is no longer size-contained since it is visible.
  requestAnimationFrame(() => {
    // Frame 2 checks:
    // The onscreen container should be visible and the switch happens immediately after layout.
    t.step(() => assert_equals(getComputedStyle(container).contain, "size layout style paint", "frame 2 onscreen"));
    t.step(() => assert_equals(getComputedStyle(offscreen_container).contain, "size layout style paint", "frame 2 offscreen"));

    requestAnimationFrame(() => {
      offscreen_container.scrollIntoView();

      // Frame 3 checks:
      // At the beginning of the frame, the onscreen container should already be visible because
      // the visibility switch should have happened after layout of the previous frame.
      t.step(() => assert_equals(getComputedStyle(container).contain, "layout style paint", "frame 3 onscreen"));
      t.step(() => assert_equals(getComputedStyle(offscreen_container).contain, "size layout style paint", "frame 3 offscreen"));

      requestAnimationFrame(() => {
        // Frame 4 checks:
        t.step(() => assert_equals(getComputedStyle(offscreen_container).contain, "size layout style paint", "frame 4"));

        requestAnimationFrame(() => {
          // Frame 5 checks:
          t.step(() => assert_equals(getComputedStyle(offscreen_container).contain, "layout style paint", "frame 4"));
          t.done();
        });
      });
    });
  });
}, "content-visibility: auto adds contain: size layout style paint;");

test(() => {
  setUp(container);
  container.classList.add("auto");
  container.style = "contain: style;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "contain: style layout;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
}, "content-visibility: auto adds contain: size layout style paint, can't be overridden");

test(() => {
  setUp(container);
  container.classList.add("auto");
  container.style = "contain: paint;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "contain: strict;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
  container.style = "contain: content;";
  assert_equals(getComputedStyle(container).contain, "size layout style paint");
}, "content-visibility keeps all containment even when shorthands are specified");
</script>
