chromium/third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-026-expected.txt

This is a testharness.js-based test.
PASS content-visibility: hidden adds contain: size layout style;
FAIL content-visibility: auto adds contain: size layout style paint; assert_equals: expected "size layout style paint" but got "layout style paint"
PASS content-visibility: auto adds contain: size layout style paint, can't be overridden
PASS content-visibility keeps all containment even when shorthands are specified
Harness: the test ran to completion.
